+++ src/main/kotlin/no/nav/medlemskap/regler/funksjoner/AdresseFunksjoner.kt
@@ -10,7 +10,7 @@ import java.time.LocalDate
 object AdresseFunksjoner {
 
 
-    fun String.landkodeErNorsk() = this == "NOR"
+
 
     fun Adresse.adressensPeriodeOverlapperKontrollPerioden(kontrollPeriode: Kontrollperiode) =
             Funksjoner.periodefilter(lagInterval(Periode(this.fom, this.tom)), kontrollPeriode.tilPeriode())
