+++ src/main/kotlin/no/nav/medlemskap/regler/v1/lovvalg/ErBrukersEktefelleBosattINorgeRegel.kt
@@ -12,7 +12,6 @@ import no.nav.medlemskap.regler.common.ja
 import no.nav.medlemskap.regler.common.nei
 import no.nav.medlemskap.regler.funksjoner.AdresseFunksjoner.adresserForKontrollPeriode
 import no.nav.medlemskap.regler.funksjoner.AdresseFunksjoner.landkodeTilAdresserForKontrollPeriode
-import no.nav.medlemskap.regler.funksjoner.RelasjonFunksjoner.hentRelatertSomFinnesITPS
 
 class ErBrukersEktefelleBosattINorgeRegel(
         ytelse: Ytelse,
