+++ src/main/kotlin/no/nav/medlemskap/regler/v1/lovvalg/ErBrukersEktefelleOgBarnasMorSammePersonRegel.kt
@@ -8,8 +8,6 @@ import no.nav.medlemskap.regler.common.RegelId
 import no.nav.medlemskap.regler.common.Resultat
 import no.nav.medlemskap.regler.common.ja
 import no.nav.medlemskap.regler.common.nei
-import no.nav.medlemskap.regler.funksjoner.RelasjonFunksjoner.hentBarnSomFinnesITPS
-import no.nav.medlemskap.regler.funksjoner.RelasjonFunksjoner.hentFnrTilBarnUnder25
 
 class ErBrukersEktefelleOgBarnasMorSammePersonRegel(
         ytelse: Ytelse,
