+++ src/main/kotlin/no/nav/medlemskap/regler/v1/lovvalg/HarBrukerBarnSomErFolkeregistrertRegel.kt
@@ -5,8 +5,6 @@ import no.nav.medlemskap.domene.barn.DataOmBarn
 import no.nav.medlemskap.regler.common.*
 import no.nav.medlemskap.regler.funksjoner.AdresseFunksjoner.adresserForKontrollPeriode
 import no.nav.medlemskap.regler.funksjoner.AdresseFunksjoner.landkodeTilAdresserForKontrollPeriode
-import no.nav.medlemskap.regler.funksjoner.RelasjonFunksjoner.hentBarnSomFinnesITPS
-import no.nav.medlemskap.regler.funksjoner.RelasjonFunksjoner.hentFnrTilBarnUnder25
 
 class HarBrukerBarnSomErFolkeregistrertRegel(
         ytelse: Ytelse,
