+++ src/main/kotlin/no/nav/medlemskap/routes/CreateDatagrunnlag.kt
@@ -4,7 +4,6 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.async
 import kotlinx.coroutines.coroutineScope
 import mu.KotlinLogging
-import no.bekk.bekkopen.person.FodselsnummerValidator
 import no.nav.medlemskap.clients.Services
 import no.nav.medlemskap.common.ytelseCounter
 import no.nav.medlemskap.domene.*
@@ -13,14 +12,11 @@ import no.nav.medlemskap.domene.barn.DataOmBarn
 import no.nav.medlemskap.domene.barn.PersonhistorikkBarn
 import no.nav.medlemskap.domene.ektefelle.DataOmEktefelle
 import no.nav.medlemskap.domene.ektefelle.PersonhistorikkEktefelle
-import no.nav.medlemskap.regler.funksjoner.RelasjonFunksjoner
-import no.nav.medlemskap.regler.funksjoner.RelasjonFunksjoner.filtrerBarnUnder25Aar
-import java.util.*
-import javax.xml.ws.soap.SOAPFaultException
-
+import no.nav.medlemskap.regler.funksjoner.ArbeidsforholdFunksjoner.fraOgMedDatoForArbeidsforhold
+import no.nav.medlemskap.regler.funksjoner.RelasjonFunksjoner.hentFnrTilBarn
+import no.nav.medlemskap.regler.funksjoner.RelasjonFunksjoner.hentFnrTilEktefelle
 
 private val logger = KotlinLogging.logger { }
-
 private val secureLogger = KotlinLogging.logger("tjenestekall")
 
 suspend fun defaultCreateDatagrunnlag(
@@ -41,16 +37,12 @@ suspend fun defaultCreateDatagrunnlag(
     val arbeidsforholdRequest = async { services.aaRegService.hentArbeidsforhold(fnr, callId, fraOgMedDatoForArbeidsforhold(periode), periode.tom) }
     val journalPosterRequest = async { services.safService.hentJournaldata(fnr, callId) }
     val gosysOppgaver = async { services.oppgaveService.hentOppgaver(aktorIder, callId) }
-  //  val personhistorikkForFamilie = hentPersonhistorikkForFamilieAsync(personHistorikkFraPdl, services, periode)
-
 
     val fnrTilBarn = hentFnrTilBarn(personHistorikkFraPdl.familierelasjoner)
     dataOmBrukersBarn = if (fnr.isNotEmpty()) hentDataOmBarn(fnrTilBarn, services, callId) else null
 
-
     val fnrTilEktefelle = hentFnrTilEktefelle(personHistorikkFraPdl)
     dataOmEktefelle = if (fnrTilEktefelle.isNullOrEmpty()) hentDataOmEktefelle(fnrTilEktefelle, services, callId, periode) else null
-
 
     val medlemskap = medlemskapsunntakRequest.await()
     val arbeidsforhold = arbeidsforholdRequest.await()
@@ -75,11 +67,7 @@ suspend fun defaultCreateDatagrunnlag(
 }
 
 suspend fun hentDataOmBarn(fnrBarn: List<String>, services: Services, callId: String): List<DataOmBarn> {
-
-    return fnrBarn
-            .map { DataOmBarn(
-                    personhistorikkBarn = hentPersonHistorikkForBarn(it, services, callId))
-            }
+    return fnrBarn.map { DataOmBarn(personhistorikkBarn = hentPersonHistorikkForBarn(it, services, callId)) }
 }
 
 private suspend fun CoroutineScope.hentDataOmEktefelle(fnrTilEktefelle: String?, services: Services, callId: String, periode: InputPeriode): DataOmEktefelle? {
@@ -109,28 +97,7 @@ private suspend fun hentPersonhistorikkFraPdl(services: Services, fnr: String, c
     return services.pdlService.hentPersonHistorikk(fnr, callId)
 }
 
-private fun hentFnrTilEktefelle(personHistorikkFraPdl: Personhistorikk?): String? {
-    val fnrTilEktefelle =
-            personHistorikkFraPdl?.sivilstand
-                    ?.filter { it.relatertVedSivilstand != null }
-                    ?.filter { FodselsnummerValidator.isValid(it.relatertVedSivilstand) }
-                    ?.filter { it.type == Sivilstandstype.GIFT || it.type == Sivilstandstype.REGISTRERT_PARTNER }
-                    ?.map { it.relatertVedSivilstand }
-                    ?.lastOrNull()
-    return fnrTilEktefelle
-}
-
-fun Familierelasjon.erBarn() = this.relatertPersonsRolle == Familierelasjonsrolle.BARN
-
-fun hentFnrTilBarn(familierelasjoner: List<Familierelasjon>): List<String> {
-    return familierelasjoner
-            .filter { FodselsnummerValidator.isValid(it.relatertPersonsIdent) }
-            .filter { it.erBarn() }
-            .filter { it.relatertPersonsIdent.filtrerBarnUnder25Aar()}
-            .map { it.relatertPersonsIdent }
-}
 
 
-private fun fraOgMedDatoForArbeidsforhold(periode: InputPeriode) = periode.fom.minusYears(1).minusDays(1)
